Fixed ADC Value Retrieval
Corrected error in ratiometric scaling where the result was off by a factor of two. The problem was due to failing to recognize that since it's a differential ADC, one bit of the resolution is used to provide the sign of the value - essentially halving the resolution.

+++ scripts/i2c_utility.py
@@ -71,7 +71,7 @@ def get_ADC_value(bus, addr, channel):
     if (sign == 1):
         val = (val ^ 0x3ff) + 1  # compute 2s complement for 12 bit val
     # Convert val to a ratiomerical ADC reading
-    return float(val) / float(4095)
+    return float(val) / float(2047)
 
 
 def IO_expander_output(bus, addr, bank, mask):
